refactor(firecrawl): separate client tools and settings
